extracted sidebar from wrapper
